Move components into dedicated folder
