+++ admin-dev/themes/new-theme/js/components/modal/modal-container.ts
